--- src/components/map/index.ts
@@ -1,1 +1,0 @@
-export * from './map-type';
